[LPF-569]: Test Failures Discrepancy: Terminal vs. IntelliJ Builds - fix integration tests
