+++ src/it/java/uk/gov/laa/gpfd/integration/GetReportsIT.java
@@ -2,10 +2,11 @@ package uk.gov.laa.gpfd.integration;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,10 +20,11 @@ import uk.gov.laa.gpfd.utils.FileUtils;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 
+@SpringBootTest
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @TestPropertySource(locations = "classpath:application-test.yml")
-@SpringBootTest
 class GetReportsIT {
 
     @Autowired
@@ -31,7 +33,7 @@ class GetReportsIT {
     @Autowired
     private JdbcTemplate writeJdbcTemplate;
 
-    @BeforeEach
+    @BeforeAll
     void setupDatabase() {
         String sqlSchema = FileUtils.readResourceToString("gpfd_schema.sql");
         String sqlData = FileUtils.readResourceToString("gpfd_data.sql");
@@ -40,7 +42,7 @@ class GetReportsIT {
         writeJdbcTemplate.execute(sqlData);
     }
 
-    @AfterEach
+    @AfterAll
     void resetDatabase() {
         writeJdbcTemplate.execute("DROP TABLE IF EXISTS GPFD.REPORT_TRACKING");
         writeJdbcTemplate.execute("DROP TABLE IF EXISTS GPFD.CSV_TO_SQL_MAPPING_TABLE");
